backend/requirements.txt
fastapi
uvicorn
openai
langchain
tiktoken
pinecone-client
python-dotenv